Keep tree from freezing when revisiting parents
Bugfix: For some reason, spacetree tries to fetch grandchildren (with n ajax queries)
when clicking a node whose children are already loaded. When this happens, level is 0.
These ajax queries will be dropped, since we we use @sequence now to ignore all but the latest.
This caused the spacetree to sit waiting in busy mode for them forever, freezing up the tree.

Also ignoring clicks on nodes that were just clicked. No need for the extra request.